refactor: Consolidate shared types and remove duplicate definitions
